Custom Weave & Repl Service CIDR bits
This allows a module user to specify a custom CIDR range for weave to use or the replicated service. It was discovered many moons ago that if the default CIDR range is already in use, you can break k8s in new and fun ways. Changes to the ptfe installer previously made, this change exposes this feature in the module to end users.