Fix pinning release sequence (#75)
* Pin release sequence

Need to pass a root module variable down to the `cloud-init` module so
that a user can pin the release sequence.

* Added documentation and a default value

Added documentation for the `release_sequence` input and arranged it so
it is alphabetical in the files. Also added a default value of `latest`
which installs the latest version of Terraform Enterprise.

* Updated with make docs

Fixed a symlink that was using an absolute path to use a relative path
instead.