Fix MIG max_unavailable_fixed for 4-zone deployment
Changed max_unavailable_fixed from 3 to 4 to satisfy GCP requirement
that this value must be 0 or >= number of zones for regional MIGs.

Error: Invalid value for field 'resource.updatePolicy.maxUnavailable.fixed': '3'.
Fixed updatePolicy.maxUnavailable for regional managed instance group has
to be either 0 or at least equal to the number of zones.

The test deployment uses 4 zones (us-central1-a/b/c/f), so the value
must be at least 4.

Refs: IND-5862